+++ src/r3d_surface_shader.c
@@ -73,7 +73,7 @@ R3D_SurfaceShader* R3D_LoadSurfaceShader(const char* filePath)
 
 R3D_SurfaceShader* R3D_LoadSurfaceShaderFromMemory(const char* code)
 {
-    int userCodeLen = strlen(code);
+    size_t userCodeLen = strlen(code);
     if (userCodeLen > R3D_MAX_SHADER_CODE_LENGTH) {
         R3D_TRACELOG(LOG_ERROR, "Failed to load surface shader; User code too long");
         return NULL;
@@ -190,7 +190,7 @@ R3D_SurfaceShader* R3D_LoadSurfaceShaderFromMemory(const char* code)
     *outPtr = '\0';
 
     // Copy transformed code to shader structure
-    int finalLen = strlen(output);
+    size_t finalLen = strlen(output);
     if (finalLen > R3D_MAX_SHADER_CODE_LENGTH) {
         R3D_TRACELOG(LOG_ERROR, "Failed to load surface shader; Transformed code too long");
         RL_FREE(output);
